// mobile/scripts/collect_data/parser/model/enumModel.ts
export const enum AccessibilityEventType {
    TYPE_VIEW_FOCUSED = 'TYPE_VIEW_FOCUSED',
    TYPE_VIEW_SCROLLED = 'TYPE_VIEW_SCROLLED',
}

export const enum TypeData {
    TEXT_ROW = 'com.whatsapp:id/conversation_text_row',
    DATE_DIVIDER = 'com.whatsapp:id/conversation_row_date_divider',
    CONTACT_NAME = 'com.whatsapp:id/conversation_contact_name'
}

export const enum TypeGenericViewIdData {
    HOUR = 'com.whatsapp:id/date',
    VIEWED = 'com.whatsapp:id/status',
}

export const enum TypeTextRowViewIdData {
    TEXT = 'com.whatsapp:id/message_text'
}

export const enum DataRole {
    USER = 'user',
    CONTACT = 'contact'
}